+++ changes.md
@@ -1,4 +1,12 @@
 # chesseval Changes
 
+2026-JUN-16
+* Added engine analysis depth + time to output figure
+* Algorithm efficiency improvements
+* Updated graph scaling
+* added command line args for analysis depth/time
+* added command line args for disabling plots
+* You can now import games from external PGN files
+
 2026-MAY-11
 * First release